[KNOWAGE-6313][KNOWAGE-6314]
KNOWAGE-6313 - https://production.eng.it/jira/browse/KNOWAGE-6313
Still not sure about this one (when it happens, what is the exact problem). I changed it so the tree is collapsed by default so the icon shouldn't show expanded status at first. I also fixed the progress bar not showing at the top, so now the user knows something is happening when he tries to expand the node.

KNOWAGE-6314 - https://production.eng.it/jira/browse/KNOWAGE-6314
Still can't reproduce this, but I think I found a possible problem inside the code. I used node id for buttonVisible/icons, but it might be possible that the content node and word node have the same id, so I added prefix and assigned it to the key property and use that one instead.